memory: move mongo chat history into the separate module
