+++ memory/mongo/mongo_chat_history_options.go
@@ -1,4 +1,4 @@
-package memory
+package mongo
 
 import (
 	"errors"
@@ -14,10 +14,10 @@ var (
 	errMongoInvalidSessionID = errors.New("invalid mongo session id option")
 )
 
-type MongoDBChatMessageHistoryOption func(m *MongoDBChatMessageHistory)
+type ChatMessageHistoryOption func(m *ChatMessageHistory)
 
-func applyMongoDBChatOptions(options ...MongoDBChatMessageHistoryOption) (*MongoDBChatMessageHistory, error) {
-	h := &MongoDBChatMessageHistory{
+func applyMongoDBChatOptions(options ...ChatMessageHistoryOption) (*ChatMessageHistory, error) {
+	h := &ChatMessageHistory{
 		databaseName:   mongoDefaultDBName,
 		collectionName: mongoDefaultCollectionName,
 	}
@@ -37,30 +37,30 @@ func applyMongoDBChatOptions(options ...MongoDBChatMessageHistoryOption) (*Mongo
 }
 
 // WithConnectionURL is an option for specifying the MongoDB connection URL. Must be set.
-func WithConnectionURL(connectionURL string) MongoDBChatMessageHistoryOption {
-	return func(p *MongoDBChatMessageHistory) {
+func WithConnectionURL(connectionURL string) ChatMessageHistoryOption {
+	return func(p *ChatMessageHistory) {
 		p.url = connectionURL
 	}
 }
 
 // WithSessionID is an arbitrary key that is used to store the messages of a single chat session,
 // like user name, email, chat id etc. Must be set.
-func WithSessionID(sessionID string) MongoDBChatMessageHistoryOption {
-	return func(p *MongoDBChatMessageHistory) {
+func WithSessionID(sessionID string) ChatMessageHistoryOption {
+	return func(p *ChatMessageHistory) {
 		p.sessionID = sessionID
 	}
 }
 
 // WithCollectionName is an option for specifying the collection name.
-func WithCollectionName(name string) MongoDBChatMessageHistoryOption {
-	return func(p *MongoDBChatMessageHistory) {
+func WithCollectionName(name string) ChatMessageHistoryOption {
+	return func(p *ChatMessageHistory) {
 		p.collectionName = name
 	}
 }
 
 // WithDataBaseName is an option for specifying the database name.
-func WithDataBaseName(name string) MongoDBChatMessageHistoryOption {
-	return func(p *MongoDBChatMessageHistory) {
+func WithDataBaseName(name string) ChatMessageHistoryOption {
+	return func(p *ChatMessageHistory) {
 		p.databaseName = name
 	}
 }
