+++ memory/mongo/mongo_chat_history_test.go
@@ -1,4 +1,4 @@
-package memory
+package mongo
 
 import (
 	"context"
